Support re-edits
