+++ brainframe_qt/ui/resources/i18n/brainframe_zh.ts
@@ -1284,22 +1284,22 @@ Please recheck the entered server address.</source>
 <context>
     <name>TaskConfiguration</name>
     <message>
-        <location filename="../../dialogs/task_configuration/task_configuration.py" line="180"/>
+        <location filename="../../dialogs/task_configuration/task_configuration.py" line="186"/>
         <source>Add points until done, then press &quot;Confirm&quot; button</source>
         <translation>添加点直到完成，然后按“确认”按钮</translation>
     </message>
     <message>
-        <location filename="../../dialogs/task_configuration/task_configuration.py" line="216"/>
+        <location filename="../../dialogs/task_configuration/task_configuration.py" line="224"/>
         <source>Item Name Already Exists</source>
         <translation>名称已存在</translation>
     </message>
     <message>
-        <location filename="../../dialogs/task_configuration/task_configuration.py" line="217"/>
+        <location filename="../../dialogs/task_configuration/task_configuration.py" line="225"/>
         <source>Item {} already exists in Stream</source>
         <translation>项目{}已存在于视频流中</translation>
     </message>
     <message>
-        <location filename="../../dialogs/task_configuration/task_configuration.py" line="220"/>
+        <location filename="../../dialogs/task_configuration/task_configuration.py" line="228"/>
         <source>Please use another name.</source>
         <translation>请使用另一个名称。</translation>
     </message>
